pyproject.toml
[tool.poetry]
name = "chess-yay"
version = "0.1.0"
description = ""
authors = ["ShanThatos <shanthkoka@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.14.1"
chess = "^1.10.0"
python-dotenv = "^1.0.1"
psutil = "^5.9.8"
uvicorn = {extras = ["standard"], version = "^0.28.0"}
fastapi = "^0.110.0"
jinja2 = "^3.1.3"
itsdangerous = "^2.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
